fix safari by polyfilling customized built-in elements
Even though the devtools are made by the chromium team for chromium-based browsers, it also works in Firefox. It doesn't work in Safari because it depends on customized built-in elements.

https://github.com/ChromeDevTools/devtools-frontend/blob/e6256afd3f8f181453984b407fe63c50de9559f5/front_end/ui/legacy/UIUtils.ts#L1230-L1233

The latest version of the devtools does not use customized built-in elements. It may work in Safari but the timeline-viewer project hasn't been updated since April 2024. Updating timeline-viewer as a non-googler is not trivial because it uses several undocumented APIs.